Created helper folder for test util to be more clear, created test_base_region.py
