+++ tests/devices/unit_tests/electron_analyser/conftest.py
@@ -30,7 +30,7 @@ from dodal.devices.electron_analyser.vgscienta import (
     VGScientaAnalyserDriverIO,
     VGScientaSequence,
 )
-from tests.devices.unit_tests.electron_analyser.util import (
+from tests.devices.unit_tests.electron_analyser.helpers import (
     get_test_sequence,
 )
