+++ tests/devices/unit_tests/electron_analyser/specs/test_driver_io.py
@@ -20,7 +20,7 @@ from dodal.devices.electron_analyser.specs import (
     SpecsAnalyserDriverIO,
     SpecsRegion,
 )
-from tests.devices.unit_tests.electron_analyser.util import (
+from tests.devices.unit_tests.electron_analyser.helpers import (
     TEST_SEQUENCE_REGION_NAMES,
     create_analyser_device,
 )
