+++ tests/devices/unit_tests/electron_analyser/specs/test_region.py
@@ -5,29 +5,20 @@ import pytest
 from dodal.common.data_util import load_json_file_to_class
 from dodal.devices.b07 import LensMode, PsuMode
 from dodal.devices.electron_analyser import EnergyMode
-from dodal.devices.electron_analyser.abstract.base_region import TAbstractBaseRegion
 from dodal.devices.electron_analyser.specs import (
     AcquisitionMode,
-    SpecsRegion,
     SpecsSequence,
 )
-from tests.devices.unit_tests.electron_analyser.util import (
-    TEST_SPECS_SEQUENCE,
+from tests.devices.unit_tests.electron_analyser.helpers import (
     assert_region_has_expected_values,
-    assert_region_kinetic_and_binding_energy,
+    get_test_sequence,
 )
 
 
 @pytest.fixture
 def sequence() -> SpecsSequence[LensMode, PsuMode]:
-    return load_json_file_to_class(
-        SpecsSequence[LensMode, PsuMode], TEST_SPECS_SEQUENCE
-    )
-
-
-@pytest.fixture
-def expected_region_class() -> type[SpecsRegion[LensMode, PsuMode]]:
-    return SpecsRegion[LensMode, PsuMode]
+    seq = SpecsSequence[LensMode, PsuMode]
+    return load_json_file_to_class(seq, get_test_sequence(seq))
 
 
 @pytest.fixture
@@ -74,45 +65,13 @@ def expected_region_values() -> list[dict[str, Any]]:
     ]
 
 
-def test_sequence_get_expected_region_from_name(
-    sequence: SpecsSequence[LensMode, PsuMode], expected_region_names: list[str]
-) -> None:
-    for name in expected_region_names:
-        assert sequence.get_region_by_name(name) is not None
-    assert sequence.get_region_by_name("region name should not be in sequence") is None
-
-
-def test_sequence_get_expected_region_type(
-    sequence: SpecsSequence[LensMode, PsuMode],
-    expected_region_class: type[TAbstractBaseRegion],
-) -> None:
-    regions = sequence.regions
-    enabled_regions = sequence.get_enabled_regions()
-    assert isinstance(regions, list) and all(
-        isinstance(r, expected_region_class) for r in regions
-    )
-    assert isinstance(enabled_regions, list) and all(
-        isinstance(r, expected_region_class) for r in enabled_regions
-    )
-
-
-def test_sequence_get_expected_region_names(
-    sequence: SpecsSequence[LensMode, PsuMode], expected_region_names: list[str]
-) -> None:
-    assert sequence.get_region_names() == expected_region_names
-
-
 def test_sequence_get_expected_enabled_region_names(
-    sequence: SpecsSequence[LensMode, PsuMode], expected_enabled_region_names: list[str]
+    sequence: SpecsSequence[LensMode, PsuMode],
+    expected_enabled_region_names: list[str],
 ) -> None:
     assert sequence.get_enabled_region_names() == expected_enabled_region_names
-
-
-def test_region_kinetic_and_binding_energy(
-    sequence: SpecsSequence[LensMode, PsuMode],
-) -> None:
-    for r in sequence.regions:
-        assert_region_kinetic_and_binding_energy(r)
+    for i, region in enumerate(sequence.get_enabled_regions()):
+        assert region.name == expected_enabled_region_names[i]
 
 
 def test_file_loads_into_class_with_expected_values(
